renamed module and it's respective references
