+++ lib/src/features/sectors/screen/sector_details/sector_details_screen_content.dart
